+++ dataduct/utils/cli.py
@@ -81,7 +81,7 @@ def open_sql_shell(database_type, host_alias=None, **kwargs):
     from dataduct.config import Config
     config = Config()
     if database_type == 'redshift':
-        open_shell.open_psql_shell()
+        open_shell.open_psql_shell(config.redshift)
     else:
         assert config.mysql.get(host_alias), \
             'host_alias "{}" does not exist in config'.format(host_alias)
@@ -123,13 +123,22 @@ pipeline_run_options = ArgumentParser(
     description='Specify actions related to running pipelines',
     add_help=False
 )
-pipeline_run_options.add_argument(
+
+group = pipeline_run_options.add_mutually_exclusive_group(required=False)
+group.add_argument(
     '-f',
     '--force',
     action='store_true',
-    default=False,
-    help='Destroy previous versions of this pipeline, if they exist',
+    help='Destroy previous versions of this pipeline, if they exist'
 )
+group.add_argument(
+    '-nf',
+    '--no-force',
+    action='store_false',
+    help='Do not destroy previous versions of this pipeline, if they exist'
+)
+pipeline_run_options.set_defaults(force=True)
+
 pipeline_run_options.add_argument(
     '-t',
     '--time_delta',